Vault is now its own class
